Upd btn action

+++ resources/views/content/posts/index.blade.php
@@ -45,12 +45,10 @@
                         </td>
                         <td>{{ $post->created_at }}</td>
 
-                        <td style="width: 110px;text-align: center">
-                            <div class="btn-group">
-                                <a href="{{ route('posts.show', $post->slug) }}" target="_blank" class="btn btn-xs btn-success"><i class="fa fa-eye"></i></a>
-                                <a href="{{ route('control.posts.edit', $post) }}" class="btn btn-xs btn-warning"><i class="fa fa-edit"></i></a>
-                                <a href="#" data-url="{{ route('control.posts.destroy', $post) }}" class="btn btn-xs btn-danger js-action-form" data-method="DELETE"><i class="fa fa-trash"></i></a>
-                            </div>
+                        <td class="wh-center btn-media">
+                            <a href="{{ route('posts.show', $post->slug) }}" target="_blank" class="btn btn-xs btn-success"><i class="fa fa-eye"></i></a>
+                            <a href="{{ route('control.posts.edit', $post) }}" class="btn btn-xs btn-warning"><i class="fa fa-edit"></i></a>
+                            <a href="#" data-url="{{ route('control.posts.destroy', $post) }}" class="btn btn-xs btn-danger js-action-form" data-method="DELETE" data-confirm="{{ trans('lte::main.Delete') }}?"><i class="fa fa-trash"></i></a>
                         </td>
                     </tr>
                     @endforeach
